Use typescript for API routes
