+++ lib/auth0.ts
@@ -1,5 +1,6 @@
 import {initAuth0} from '@auth0/nextjs-auth0'
 import ms from 'ms'
+import {NextApiRequest} from 'next'
 
 const cookieLifetime = ms('30 days') / 1000
 const httpTimeout = ms('10s')
@@ -8,7 +9,7 @@ const scope = 'openid profile id_token'
 /**
  * Read the origin from the request to configure the redirect urls
  */
-export default (req) => {
+export default (req: NextApiRequest) => {
   const host = req.headers.host
   const protocol = /^localhost(:\d+)?$/.test(host) ? 'http:' : 'https:'
   const origin = `${protocol}//${host}`
