+++ package.json
@@ -64,7 +64,9 @@
     "@turf/line-slice": "^5.1.5",
     "@turf/nearest-point-on-line": "^6.0.2",
     "@turf/point-to-line-distance": "^6.0.0",
+    "@types/http-proxy": "^1.17.4",
     "@types/lodash": "^4.14.150",
+    "@types/ms": "^0.7.31",
     "@types/react": "^16.9.34",
     "@zeit/next-mdx": "^1.2.0",
     "axios": "^0.19.2",
@@ -148,7 +150,8 @@
     "redux-mock-store": "^1.5.4",
     "regenerator-runtime": "^0.13.5",
     "rimraf": "^3.0.2",
-    "semantic-release": "^17.0.6"
+    "semantic-release": "^17.0.6",
+    "typescript": "^3.8.3"
   },
   "browser": {
     "node-fetch": false
