+++ pages/api/logout.ts
@@ -1,11 +1,14 @@
+import {NextApiRequest, NextApiResponse} from 'next'
+
 import initAuth0 from '../../lib/auth0'
 
-export default async function callback(req, res) {
+export default async function logout(
+  req: NextApiRequest,
+  res: NextApiResponse
+) {
   try {
     const auth0 = initAuth0(req)
-    await auth0.handleCallback(req, res, {
-      redirectTo: req.query.redirectTo
-    })
+    await auth0.handleLogout(req, res)
   } catch (error) {
     console.error(error)
     res.status(error.status || 500).end(error.message)
